requirements.txt
numpy
matplotlib
cdlib
requests
python-dotenv
sklearn
mysql-connector-python
networkx
scikit-learn
folium
infomap
wurlitzer
pandas